Wippety WIP config class
